+++ tests/config/bundles.php
@@ -6,4 +6,5 @@ return [
     Symfony\Bundle\TwigBundle\TwigBundle::class => ['all' => true],
     Schvoy\BaseEntityBundle\BaseEntityBundle::class => ['all' => true],
     Schvoy\MailTemplateBundle\MailTemplateBundle::class => ['all' => true],
+    Symfony\Bundle\SecurityBundle\SecurityBundle::class => ['all' => true],
 ];